add configuration class to product-service
